backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7

# Email
email-validator>=2.1.0

